:label: add type annotation in XGBoost model
